ValBot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytesseract==0.3.4
requests==2.24.0
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
yarl==1.5.1
